Add FIRE_{GUN,MISSILE} controls to Pilot
Handle player inputs for firing guns and missiles in PlayerPilot

Add the aforementioned controls to the controls array in Ship.gd

Clean up the code a bit by removing or commenting some unused stuff, and making the code more readable. Note that  unused stuff that may be used later is commented for now.

Do not use if/else to handle boolean non-toggle inputs