Do some housekeeping and tweaks to Space.tscn/gd
Change the way the cockpit and camera work. The cockpit camera is now a child of the player's ship, and its transform is copied from the CockpitPosition node in Ship.tscn. The chase camera is now an InterpolatedCamera that "follows" the ChasePosition node in Ship.tscn.

Add background music to Space.tscn

Remove the "simple skybox" and the cubemap textures in favour of the equirectangular textures Godot wants to use for environment backgrounds

Make it easy to replace the equirectangular texture used as the background