Make ship point forwards, and rework bullet collisions
Use collision exceptions for bullets to prevent them from colliding with the shooter.

Scale ship model and its respective shield model by -1 on the Z axis. I should really just re-export the OBJs, but this is a quick and dirty solution.

Move chase camera and gun points to match the ship's new orientation.