Attempt to fix weapon firing
I don't know what went wrong, but for some reason, the weapons were not firing. It looks like part of the problem was the fact that I renamed get_refire to _get_refire (prefixed with an underscore). I had to re-connect the signal to _on_Player_weapon_fired in the editor in order to fix weapons firing.