Use a dictionary instead of an array to store radar dots
I believe this will be more efficient because the key is a pointer to the ship, and the value is a dictionary with the info about the dot like position, colour, and texture.
Modify Ship._visible_on_radar to return a texture so that individual ships can customize their appearance on radar.