Work on framework for weapon requirements
Change Bullet._fire and Bullet._unfire return values to dictionaries specifying whether or not the firing was successful, and the firing requirements.

Add max_ammo to Weapon, and rename "ammunition" to "ammo".

Pass an "inventory" dictionary to Weapon._fire instead of a "energy" value.

Add optional "charge" argument to Bullet._get_requirements, which helps calculate the energy requirements for a charging weapon.