Initial work on projectile interception
Since this is a game where almost everything is always moving, the AI needs to know how to lead its target in order to land hits on it.
Add projectile intercept calculation code to radartest2.gd.
Add "InterceptTarget" nodes to radartest2.tscn in order to test the projectile intercept calculation code, and make it follow the intercept point of the enemy ship firing its weapons at the player's ship.